FROMLIST: pinctrl: intel: Fix a glitch when updating IRQ flags on a preconfigured line
The commit af7e3eeb84e2 ("pinctrl: intel: Disable input and output buffer
when switching to GPIO") hadn't taken into account an update of the IRQ
flags scenario.

When updating the IRQ flags on the preconfigured line the ->irq_set_type()
is called again. In such case the sequential Rx buffer configuration
changes may trigger a falling or rising edge interrupt that may lead,
on some platforms, to an undesired event.

This may happen because each of intel_gpio_set_gpio_mode() and
__intel_gpio_set_direction() updates the pad configuration with a different
value of the GPIORXDIS bit. Notable, that the intel_gpio_set_gpio_mode() is
called only for the pads that are configured as an input. Due to this fact,
integrate the logic of __intel_gpio_set_direction() call into the
intel_gpio_set_gpio_mode() so that the Rx buffer won't be disabled and
immediately re-enabled.

Fixes: af7e3eeb84e2 ("pinctrl: intel: Disable input and output buffer when switching to GPIO")
Reported-by: Kane Chen <kane.chen@intel.com>
Signed-off-by: Andy Shevchenko <andriy.shevchenko@linux.intel.com>
(am from https://patchwork.ozlabs.org/project/linux-gpio/patch/20220119181915.27519-1-andriy.shevchenko@linux.intel.com/)

BUG=b:212346252,b:214315232
TEST=Verified suspend stress 100 cycles with the stylus plugged/unplugged

Signed-off-by: Grace Kao <grace.kao@intel.corp-partner.google.com>
Change-Id: I69bb27f30c235df77d535eaa1a5cf82f4b887d29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3403369
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Rajat Jain <rajatja@google.com>
Reviewed-by: David Wu <david_wu@quanta.corp-partner.google.com>
Tested-by: David Wu <david_wu@quanta.corp-partner.google.com>