UPSTREAM: media: mc: Provide a helper for setting bus_info field
The bus_info or a similar field exists in a lot of structs, yet drivers
tend to set the value of that field by themselves in a determinable way.
Thus provide a helper for doing this. To be used in subsequent patches.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit 78a171e58717685ea0eb082a24bc75bd9e20518f)

BUG=b:230028419
BUG=b:232131883
TEST=v4l2-ctl -D [decoder] should show bus_info for media device

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I54649f980c4c098292833704d351c6dcef9fa597
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3769352
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Steve Cho <stevecho@chromium.org>
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>