BACKPORT: media: v4l: ioctl: Set bus_info in v4l_querycap()
The bus_info field is set by most drivers based on the type of the device
bus as well as the name of the device. Do this in v4l_querycap() so
drivers don't need to. This keeps compatibility with non-default and silly
bus_info.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>

(cherry picked from commit f2d8b6917f3bcfb3190eb80567fea71a9b59dbd3)

Conflicts:
   drivers/media/platform/nxp/imx-jpeg/mxc-jpeg.c
   drivers/media/platform/ti/cal/cal-video.c

BUG=b:230028419
BUG=b:232131883
TEST=v4l2-ctl -D [decoder] should show bus_info for media device

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I827a4abeedbf755f10e4f2f4528894db04bde34b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3769354
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Steve Cho <stevecho@chromium.org>