FROMLIST: usb: typec: Only attempt to link USB ports if there is fwnode
The code that creates the links to the USB ports attached to
a connector inside the system assumed that the ACPI nodes
(fwnodes) always exist for the connectors, but it can not do
that.

There is no guarantee that every USB Type-C connector has
ACPI device node representing it in the ACPI tables, and
even if there are the nodes in the ACPI tables, the _STA
method in those nodes may still return 0 (which means the
device does not exist from ACPI PoW).

This fixes NULL pointer dereference that happens if the
nodes are missing.

Reported-and-tested-by: Robert Święcki <robert@swiecki.net>
Reported-by: Mikhail Gavrilov <mikhail.v.gavrilov@gmail.com>
Fixes: 730b49aac426 ("usb: typec: port-mapper: Convert to the component framework")
Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
(am from https://patchwork.kernel.org/patch/12721695/)
(also found at https://lore.kernel.org/r/20220124090228.41396-2-heikki.krogerus@linux.intel.com)

BUG=b:215551841
TEST=Boot reven on X1 Carbon gen6, verify no kernel panic

Signed-off-by: Nicholas Bishop <nicholasbishop@google.com>
Change-Id: Id9bed0ad2c690a7c98ce5b7af72a005cfee4b0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3412692
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Commit-Queue: Prashant Malani <pmalani@chromium.org>