CHROMIUM: coiommu: allow the coiommu endpoint drivers to bind again
The BDF of an endpoint is already checked at the beginning of the setup
thus the endpoint which can pass this is confirmed as on top of coiommu.
For such endpoint that has already been configured is actually by the
coiommu, so let its driver continue to work instead of stopping it.

BUG=b:218215844, b:188481989
TEST=boot chrome os vm with coiommu enabled. There is audio when playing
youtube.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.corp-partner.google.com>
Change-Id: I117c61cd9084a53aa0fe2b6d0bf7d6a13beced30
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3443138
Reviewed-by: David Stevens <stevensd@chromium.org>
Commit-Queue: David Stevens <stevensd@chromium.org>