CHROMIUM: ASoC: amd: Enable legacy audio on Chrome boards
With I2C PSP arbitration enabled, we are seeing errors starting SOF.
This is because the I2C arbitration and SOF are using the same PSP
mailbox register without a mutex. So the two drivers are racing.

Change acp-config flag to use legacy audio to fix this issue.

BUG=b:220886998
TEST=Verified audio function on the legacy path.

Change-Id: Ie7c957f1e8067e18a8f526cc1253a4d256d01a4f
Signed-off-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3488249
Reviewed-by: Karthikeyan Ramasubramanian <kramasub@google.com>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Raul Rangel <rrangel@chromium.org>