soc/google/cpif: Refine PCIe IRQs affinity setting
Not like mailbox, A TPMON item for PCIe controls not only PktProc IRQ
affinity but base MSI IRQ. The number of queue can be "1" even if
PktProc is disabled. But actually we don't need to consider the case
in TPMON since base IRQ will be set as default.

Check for each queue only when PktProc is enabled to improve
readability.

Change-Id: Ica4204e8155ae2a9c3a37ac9e93cac202797f5f5
Signed-off-by: Dongseok Yi <dseok.yi@samsung.com>