CHROMIUM: arm64: dts: rockchip: Assign gpu sustainable-power
Providing a power-model for use with the Intelligent Power Allocator
(for thermal management) doesn't make much sense without providing its
sustainable power (the amount of power we can expect to be able to
dissipate) -- otherwise, IPA is too aggressive in trying to throttle far
before we get close to the 75C temperature threshold.

According to the docs, this can be a bit of an estimation game, and as
long as we're close, the PID algorithm can usually take care of things.
Other advice suggests getting a ballpark figure by figuring out the OPP
where we can sustainable run without throttling, and calculate its power
usage (according to the provided model).

So far, I find that we can run heavy graphics benchmarks even at the
highest OPP (1.075V/800MHz), but in order to allow IPA a bit of headroom
for throttling if it really needs to, I try the second-highest OPP
(0.925V/600MHz). I approximate the power at that OPP with:

  power = static power + dynamic power
        = static-coefficient * (V1)^3 + dynamic-coeff * (V2)^2 * f
        = (411522 * 10^-6 W/V^3) * (0.8V)^3 +
          (977 * 10^-12 W)/(Hz * V^2) * (0.925V)^2 * (600 * 10^6 Hz)
        = 0.8 W + 0.5 W

Tests show that even values of 1 W are enough to avoid throttling on
benchmark workloads that used to run fine on kernel 4.4, with
temperatures remaining stably below 60-65C.

BUG=b:202316516
TEST=graphics.GFXBench, graphics_GLBench

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: I05f03f7dc0c43bfbdf4bc33f8f42f5293e74c189
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3508443
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>
Commit-Queue: Chen-Yu Tsai <wenst@chromium.org>