FROMLIST: cpufreq: intel_pstate: Use firmware default EPP
For some specific platforms (E.g. AlderLake) the balance performance
EPP is updated from the hard coded value in the driver. This acts as
the default and balance_performance EPP. The purpose of this EPP
update is to reach maximum 1 core turbo frequency (when possible) out
of the box.

Although we can achieve the objective by using hard coded value in the
driver, there can be other EPP which can be better in terms of power.
But that will be very subjective based on platform and use cases.
This is not practical to have a per platform specific default hard coded
in the driver.

If a platform wants to specify default EPP, it can be set in the firmware.
If this EPP is not the chipset default of 0x80 (balance_perf_epp unless
driver changed it) and more performance oriented but not 0, the driver
can use this as the default and balanced_perf EPP. In this case no driver
update is required every time there is some new platform and default EPP.

If the firmware didn't update the EPP from the chipset default then
the hard coded value is used as per existing implementation.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
(am from https://patchwork.kernel.org/patch/12776982/)
(also found at
https://patchwork.kernel.org/project/linux-pm/patch/20220310224223.684007-1-srinivas.pandruvada@linux.intel.com/)

BUG=b:219785001
TEST=boot and read msr: rdmsr -a 0x774

Signed-off-by: George D Sworo <george.d.sworo@intel.com>
Change-Id: I87b5cfa3de7d8721e4fc9641122bc3b533d2a9ed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3518445
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Tested-by: Deepika Punyamurtula <deepika.punyamurtula@intel.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>