CHROMIUM: iwl7000: mvm: don't send protocol offload command
OFFLOADS_QUERY_CMD is sent only for tracing and doesn't have a special
handler.
A few months ago the flow had changed and we no longer
support commands and responses on the resume flow.

Since the command is sent only for tracing it is fair to just
remove it to reduce code complexity.

Change-Id: I66d085370525aeb77abb09a27d7cac0cef9f6f30
Signed-off-by: Haim, Dreyfuss <haim.dreyfuss@intel.com>
iwl7000-tree: 243b50956227119c82215c7f37d9acc1cafd43ae