CHROMIUM: config: borealis-vm: enable suspend support
Enabling suspend support allows the system to enter suspend state. This
change is required as a part of full guest suspension/resumption support
and allows to properly suspend the guest VM by calling suspend hooks for
given drivers/subsystems. This change is especially needed when using
pass-through devices, which drivers (and required suspend hooks) resides
on the guest.

BUG=b:194391015
TEST=Suspend borealis VM by: "echo mem > /sys/power/state".

Change-Id: I704a4b228e809856ca4aaef1a17b71fcc60fd51c
Signed-off-by: Grzegorz Jaszczyk <jaz@semihalf.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3541928
Reviewed-by: David Riley <davidriley@chromium.org>