CHROMIUM: HID: multitouch: Add lid handler for touchpad on Redrix
If user closes the lid the touchscreen gets close to the touchpad surface,
which causes interference and makes the touchpad enter noise mode.
Right after opening the lid the cursor is unresponsive because of the mode
mentioned.
To fix this issue we switch the surface off when closing the lid and on
then opening it. This way Elan's FW may perform recalibration once
the lid has been opened.

BUG=b:221094337
TEST=Compile, upload to redrix, verify the unresponsive issue is resolved

Change-Id: Id1b1f14b13d0d79e0b18a366b19e0b105fbb9447
Signed-off-by: Angela Czubak <acz@semihalf.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3578852
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>