FROMGIT: thunderbolt: Link USB4 ports to their USB Type-C connectors
Creating a symlink pointing to the correct USB Type-C
connector for the on-board USB4 ports when they are created.
The link will be created only if the firmware is able to
describe the connection between the port and its connector.

Signed-off-by: Heikki Krogerus <heikki.krogerus@linux.intel.com>
Signed-off-by: Won Chung <wonchung@google.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(cherry picked from commit 5dddb41692849594862473e4ddf57fb52e1249d6
 git://git.kernel.org/pub/scm/linux/kernel/git/westeri/thunderbolt.git next)

BUG=b:216691605 b:223428686
TEST=Compile + Update to Brya device

Change-Id: I6a3e337cc7f44d09c1083b4b57ad49687aa4527a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3575116
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Won Chung <wonchung@google.com>
Tested-by: Won Chung <wonchung@google.com>