UPSTREAM: landlock: Reduce the maximum number of layers to 16
The maximum number of nested Landlock domains is currently 64.  Because
of the following fix and to help reduce the stack size, let's reduce it
to 16.  This seems large enough for a lot of use cases (e.g. sandboxed
init service, spawning a sandboxed SSH service, in nested sandboxed
containers).  Reducing the number of nested domains may also help to
discover misuse of Landlock (e.g. creating a domain per rule).

Add and use a dedicated layer_mask_t typedef to fit with the number of
layers.  This might be useful when changing it and to keep it consistent
with the maximum number of layers.

Reviewed-by: Paul Moore <paul@paul-moore.com>
Link: https://lore.kernel.org/r/20220506161102.525323-3-mic@digikod.net
Cc: stable@vger.kernel.org
Signed-off-by: Mickaël Salaün <mic@digikod.net>
(cherry picked from commit 75c542d6c6cc48720376862d5496d51509160dfd)

BUG=b:232966959
TEST=test landlock sample

Signed-off-by: jeff xu <jeffxu@google.com>
Change-Id: I1e29154ef967361e1aa33eb797af6591ff81ba91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3736122
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Mickaël Salaün <mic@digikod.net>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>