CHROMIUM: iwl7000: mvm: remove not needed initializations
In iwl_mvm_add_sta() we're initializing the new allocated mvm_sta.
We are setting some fields to zero even though it was allocated with
kzalloc, for the case of HW restart. But in such a case we will never
get to this initializations due to the goto statement that we have in
this case. Remove these initializations.

Change-Id: I6959bf1aca74c865e3c1edbf711f5fe8fb8c0c9f
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 7b189cb138ef481decb199c969bb86f0def5e5fa