Revert "CHROMIUM: media: mtk-vcodec: Adapt stateless VP9 decoder to new final uAPI"
This reverts commit 1a14b037201726a6eb3a24bc449a2cdc3fded0d2.

With the migration of the VP9 stateless decoder uAPI to the finalized
version, the mtk-vcodec driver needs to be migrated as well. However
this part of the driver is not fully upstreamed yet, and the current
downstream version is also very different from the latest upstream
mailing list one.

Instead of trying to back out the whole driver and backport the latest
stuff, causing a lot of code churn and most likely breaking things in
the process, it is much easier to just update the driver to match the
new uAPI. The changes are the same as the ones that were made to the
upstream patches.

Upon uprev for cherry, the whole stack will be thrown away, and
hopefully upstream will already be in a usable state.

BUG=b:232131883
TEST=emerge-cherry sys-kernel/chromeos-kernel-5_10

Signed-off-by: Yunfei Dong <yunfei.dong@mediatek.com>
Change-Id: I8ffa3dc0b6cab1b02122d4e16e28530ebcfa178b