phy-exynos-usbdrd: Control S2MPU from exynos-pd
Previously S2MPU_HSI0 power management was tied to the power state of
dwc3. This does not account for situations where USB remains on during
CPU suspend. The correct way of handling it is the same as LDO with
hooks in exynos-pd. Add the same hooks for the S2MPU but post-power-on
and pre-power-off (unlike LDO) because the S2MPU can only be configured
while the power is on.

Bug: 233155304
Signed-off-by: David Brazdil <dbrazdil@google.com>
Change-Id: I8cc2197b7b510bf13ca5b658f16b581939052650
(cherry picked from commit f7c13c301bdb7010f086db3f9db26215e44876ae)