BACKPORT: KVM: x86: do not report a vCPU as preempted outside instruction boundaries
If a vCPU is outside guest mode and is scheduled out, it might be in the
process of making a memory access.  A problem occurs if another vCPU uses
the PV TLB flush feature during the period when the vCPU is scheduled
out, and a virtual address has already been translated but has not yet
been accessed, because this is equivalent to using a stale TLB entry.

To avoid this, only report a vCPU as preempted if sure that the guest
is at an instruction boundary.  A rescheduling request will be delivered
to the host physical CPU as an external interrupt, so for simplicity
consider any vmexit *not* instruction boundary except for external
interrupts.

It would in principle be okay to report the vCPU as preempted also
if it is sleeping in kvm_vcpu_block(): a TLB flush IPI will incur the
vmentry/vmexit overhead unnecessarily, and optimistic spinning is
also unlikely to succeed.  However, leave it for later because right
now kvm_vcpu_check_block() is doing memory accesses.  Even
though the TLB flush issue only applies to virtual memory address,
it's very much preferrable to be conservative.

Reported-by: Jann Horn <jannh@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

(cherry picked from commit 6cd88243c7e03845a450795e134b488fc2afb736)

Conflicts:
   arch/x86/include/asm/kvm_host.h
   arch/x86/kvm/x86.c

Backport Note:
* kvm_vcpu_stats_desc is not present in chromeos-5.10; was introduced in:
ce55c049459c ("KVM: stats: Support binary stats retrieval for a VCPU")

BUG=b:245854706
TEST=CQ

Signed-off-by: Zubin Mithra <zsm@google.com>
Change-Id: I9411de30482c6461053c2460571f7963d3afe7b8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3911315
Commit-Queue: Zubin Mithra <zsm@chromium.org>
Tested-by: Zubin Mithra <zsm@chromium.org>
Reviewed-by: Guenter Roeck <groeck@chromium.org>