CHROMIUM: mtk-vcodec: fix data_offset buffer overrun
Thanks to Nicolas Dufresne for pointing out both the bug and the fix.

When we validate that the image size does not exceed the size of the
plane buffer, we are currently neglecting to count the data_offset in
our calculation. Most drivers simple ignore data_offset, but the MTK
vcodec driver has been hacked specifically to support it for ARC++
encoding. See cr/2721869.

Note that this CL will need to be ported for 5.15 as well. We ported to
5.4 in cr/3696583.

BUG=b:235392268
TEST=Kernel builds and runs on Jacuzzi
Suggested-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Signed-off-by: Justin Green <greenjustin@google.com>

Change-Id: Id21da3c2b31f55ce9cc9b9bcc818756fbc369a8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3703616
Reviewed-by: Miguel Casas <mcasas@google.com>
Tested-by: Justin Green <greenjustin@google.com>
Commit-Queue: Justin Green <greenjustin@google.com>