soc/google/cpif: Clean out stop_napi_poll
There were 3 use cases of stop_napi_poll, but only PktProc uses the
func now.

If PktProc calls stop_napi_poll when its buffer is full, PktProc stop
cleaning up buffer and the free space could be too small. At that
time, CP would want to send many payloads that exceed free space.
Without CP CL#2956084 "To send MSI again if AP descriptor is not
freed", it seems like that CP does not send interrupts due to the lack
of buffer.

I think this fix is not required if the CP CL is applied. But I
believe it is the safer approach.

Bug: 234901578
Signed-off-by: Dongseok Yi <dseok.yi@samsung.com>
Change-Id: I356506dc4869fb38d075b81bbc3a32bf545e882f