FROMGIT: media: mediatek: vcodec: Initialize decoder parameters for each instance
The decoder parameters are stored in each instance's context data. This
needs to be initialized per-instance, but a previous fix incorrectly
changed it to only be initialized for the first opened instance. This
resulted in subsequent instances not correctly signaling the requirement
for the Requests API.

Fix this by calling the initializing function outside of the
v4l2_fh_is_singular() conditional block.

Fixes: faddaa735c20 ("media: mediatek: vcodec: Initialize decoder parameters after getting dec_capability")
Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Reviewed-by: Yunfei Dong <yunfei.dong@mediatek.com>
Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
(cherry picked from commit fe3d651627d61210c6905339e5281d3b9db75033
 https://git.linuxtv.org/media_stage.git master)

BUG=b:210678227
BUG=b:232131883
TEST=VideoDecoderTest.FlushAtEndOfStream_MultipleConcurrentDecodes
     should pass for all supported formats

Change-Id: I86f258269ea0dade773439898991e13c6fdc4800