CHROMIUM: media: intel/ipu6: make IPU DMA mapping work with VT-d enabled
Commit <00e90527534e7c08779a185ad8f9fddbd817daae> was trying to make
IOMMU bypass the IPU and the DMA remapping working in identical
mapping mode, the change has some drawbacks, it is hard to maintain
as it can not be upstreamed to community as IPU driver is not in
kernel tree. Besides, the patch has devices dependency, when trying to
enable new IPU hardware, we need add the new vendor ID into the bypass
list to support that, it is not easy to maintain.

This patch is trying to make IPU DMA work with IOMMU dynamic mapping
mode, the main update is trying to return the concentrated sg number
during nested mapping as IOMMU dma mapping APIs likely concentrates
the sgs into less ones.

BUG=None
TEST=Sanity checked basic camera functions on Redrix and Nivviks

Signed-off-by: Bingbu Cao <bingbu.cao@intel.com>
Change-Id: Id0ba5de93a1fc794ca21179558cef17b9176edbc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3746826
Commit-Queue: Sergey Senozhatsky <senozhatsky@chromium.org>
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>