CHROMIUM: arm64: dts: mt8195: enable panel hpd
After crrev/c/3743163, the eDP panel resume logic relies on a newly
added hpd_reliable delay or the hpd_gpio dt setting to wait for a
reliable hpd. Without the delay, the display driver continues the
resume process and fails to get the edid as the panel is not ready.
As a result, it assumes the panel is disconnected.

Both tomato and dojo use the GPIO 7 as the panel hpd, so we go with
the DT setting to minimize the panel resume time.

BUG=b:238991621
TEST=verify display resumes successfully in the laptop, extended, and
dock modes

Change-Id: I142a23d312ff4ba63f005d1e8832a1af3d537513
Signed-off-by: Wei-Shun Chang <weishunc@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3761743
Reviewed-by: Pin-yen Lin <treapking@chromium.org>