UPSTREAM: usb: typec: retimer: Add missing id check in match callback
The fwnode_connection_find_match() function handles two cases: named
references and graph endpoints. In the second case, the match function
passed in is called with the id to check for the match. However, the
match function for the recently added type-c retimer class assumes the
connection has already been matched (which is only true for the first
case).

The result is that with that change, all type-c nodes with graph
endpoints defer probe indefinitely, independently of having a retimer
connection or not.

Add the missing check, like is done by the type-c mux and usb role
switch code, to fix the issue.

Fixes: ddaf8d96f93b ("usb: typec: Add support for retimers")
Reviewed-by: Prashant Malani <pmalani@chromium.org>
Signed-off-by: Nícolas F. R. A. Prado <nfraprado@collabora.com>
Link: https://lore.kernel.org/r/20220725203129.1973260-1-nfraprado@collabora.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 0c34043897736a3fac259564a6323a1f045100bb)

BUG=b:208883109
TEST=None

Signed-off-by: Prashant Malani <pmalani@chromium.org>
Change-Id: I2500d5038539d3c3d5bfd660e59f8b6e9f52db5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3811482
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>