FROMGIT: vfs: function to prevent re-use of block-device-based superblocks
The function is to be called from filesystem-specific code to mark a
superblock to be ignored by superblock test and thus never re-used.  The
function also unregisters bdi if the bdi is per-superblock to avoid
collision if a new superblock is created to represent the filesystem.
generic_shutdown_super() skips unregistering bdi for a retired superlock as
it assumes retire function has already done it.

This patch adds the functionality only for the block-device-based supers,
since the primary use case of the feature is to gracefully handle force
unmount of external devices, mounted with FUSE.  This can be further
extended to cover all superblocks, if the need arises.

Signed-off-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
(cherry picked from commit 04b9407197789c81fffac52921e703cb47967d6a
 git://git.kernel.org/pub/scm/linux/kernel/git/mszeredi/fuse.git for-next)

BUG=b:226570055
TEST=CQ

Signed-off-by: Daniil Lunev <dlunev@google.com>
Change-Id: I2eb1f18550422f278f138ded7e436a608b47f855
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3807734
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Daniil Lunev <dlunev@chromium.org>