CHROMIUM: HID: multitouch: Add lid handler for touchpad on Vell
It seems that Vell might have some problems with touchpad and touchscreen
interfering with each other once the lid is closed. This in turn makes
the touchpad enter noise mode and causes the cursor to become laggy.
Use a lid handler so that we can trigger touchpad recalibration once the
lid has been opened.

BUG=b:237397069
TEST=Compile the kernel

Signed-off-by: Angela Czubak <acz@semihalf.com>
Change-Id: Ifc8aa096455b085f67b3d9e607f541b7e8310cfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3788411
Reviewed-by: Sean O'Brien <seobrien@chromium.org>