CHROMIUM: tpm: always block messages when suspended
A previous patch [1] blocks messages when the tpm itself and the controller
of the bus are suspended. The code contains an exception to this: if the
tpm is "always powered" or "firmware power managed", then the driver does
not need to send a stop message to the TPM, and messages are not blocked in
suspend.

When messages are not blocked in suspend, the tpm may try to access the bus
while it is in suspend. This generates intermittent kernel warnings on brya
platform ChromeOS devices:

  i2c_designware i2c_designware.1: Transfer while suspended

Fix by always blocking messages in suspend, even if the tpm does not
require a stop message in suspend.

[1] "CHROMIUM: tpm: block messages while suspended"

BUG=b:235553202
TEST=2 cycles of power_SuspendStress/control.bareFSI autotest
     on ChromeOS Banshee devices.
     each test cycle consists of 3000 suspend/resume cycles
     test_that --autotest_dir ~/trunk/src/third_party/autotest/files/ \
       --board=$BOARD $DUT_IP f:.*power_SuspendStress/control.bareFSI

Signed-off-by: Sven van Ashbrook <svenva@google.com>
Change-Id: I52a3023e2ed00fac5087d9f387dd7a4b2143e584
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3806460
Reviewed-by: Andrey Pronin <apronin@chromium.org>