CHROMIUM: config: reven: Add CONFIG_VMD
Necessary for finding the disks on some devices,
e.g. the HP 250 G8 (b/233895946).

Include as a built-in, rather than a module, to ensure that vmd is
available before the root is loaded. Without vmd we might not be able
to load the root.

BUG=b:233895946
TEST=install and then boot on an HP 250 G8

Change-Id: If6bc789f7743883b8847a36b48775bb09fdb822a
Signed-off-by: Ted Brandston <tbrandston@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3805402
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>