UPSTREAM: wifi: rtw89: rewrite decision on channel by entity state
We need to invoke the callback of the changed band at the first
set_channel() after every power-off. Originally, we forced the
channel to be 0 when doing power-off, and then determined things
by comparing channel with 0.

However, deciding on such things by channel might be confusing.
It's also confusing to use this kind of decision when we consider
multiple channels in the follow-up patches. So, another flag,
entity_active, is added ahead to HAL to deal with this.

Besides, we also need to check if entity is active when we set
TX power.

Signed-off-by: Zong-Zhe Yang <kevin_yang@realtek.com>
Signed-off-by: Ping-Ke Shih <pkshih@realtek.com>
Signed-off-by: Kalle Valo <kvalo@kernel.org>
Link: https://lore.kernel.org/r/20220809104952.61355-2-pkshih@realtek.com
(cherry picked from commit 967439c7a2a60923f9b4c8ac925d368af16fb165)

BUG=b:246879086
TEST=suite:wifi_matfunc on Taniks

Change-Id: Id6ebeb25b6bca8e9584d472adb8658a148f99b08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3930448
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Arowa Suliman <arowa@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Arowa Suliman <arowa@chromium.org>