CHROMIUM: platform/chrome: disable magnetometer for reef/electro
Compass feature is not needed on reef/electro devices, so disable it
by applying a workaround in the kernel cros_ec_sensorhub driver.
It is required in case running without updated firmware, as
the information about sensors available in the system is
obtained from EC in runtime.

BUG=b:249419684
TEST=none
UPSTREAM-TASK=b:253013848

Signed-off-by: Marcin Wojtas <mwojtas@google.com>
Change-Id: I1d2137bc78a6a3938f93d3144f37fc2a7cf70593
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3936629
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>
Reviewed-by: Matt Davis <mattedavis@google.com>
Reviewed-by: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>