CHROMIUM: iwl7000: mvm: remove useless code
Setting the station to -EBUSY was originally done under
this lock, and the comment still refers to it. But this
no longer happens because that was removed when DQA was
removed. Remove the leftover code as well.

Change-Id: Ie2c2ff6fd7c6e3ebf5b736de350dc15515970792
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
tested: iil_jenkins <EC.GER.UNIX.IIL.JENKINS@INTEL.COM>
iwl7000-tree: 523b554d2ba8fdcb756522123504277dc5b8ad68