BACKPORT: UPSTREAM: drm/mediatek: Add MT8195 Embedded DisplayPort driver
This patch adds a embedded displayport driver for the MediaTek mt8195
SoC.

It supports the MT8195, the embedded DisplayPort units. It offers
DisplayPort 1.4 with up to 4 lanes.

The driver creates a child device for the phy. The child device will
never exist without the parent being active. As they are sharing a
register range, the parent passes a regmap pointer to the child so that
both can work with the same register range. The phy driver sets device
data that is read by the parent to get the phy device that can be used
to control the phy properties.

This driver is based on an initial version by
Jitao shi <jitao.shi@mediatek.com>

BACKPORT Reason: Some customers are using the old version of dp_driver, the bug(BUG=b:230575591) we fixed in the new version of upstreaming dp_driver needs to backport back to the old version to support customers.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Guillaume Ranquet <granquet@baylibre.com>
Signed-off-by: Bo-Chen Chen <rex-bc.chen@mediatek.com>
Reviewed-by: CK Hu <ck.hu@mediatek.com>
Tested-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Reviewed-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com>
Signed-off-by: Dmitry Osipenko <dmitry.osipenko@collabora.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220901044149.16782-4-rex-bc.chen@mediatek.com
(cherry picked from commit f70ac097a2cf5d4b67b2c1bbb73196c573ffcb7b)

BUG=b:229708673
TEST=emerge-cherry sys-kernel/chromeos-kernel-5_10

Signed-off-by: Liangxu Xu <liangxu.xu@mediatek.corp-partner.google.com>
Signed-off-by: Xinlei Lee <Xinlei.Lee@mediatek.corp-partner.google.com>
Change-Id: I3096849f2c7522ddb0e2362991351b409223e42e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3944389
Commit-Queue: Wei-Shun Chang <weishunc@chromium.org>
Reviewed-by: Wei-Shun Chang <weishunc@chromium.org>