CHROMIUM: usb: Add NO_LPM quirk to WA compatibility issue on Type-C port with PS8815 retimer
A compatibility issue of Type-C port with PS8815 retimer is identified,
the USB 3.2 Gen2 external SSD device can't be enumerated successfully
and system fail to recognize this removable media in Files apps.

By tuning USB 3.2 receiver magin, the revised NPHY FW likely to resolve this issue.
This patch is considered as temporary WA before the solution is available.
Once the FW is validated and reviewed, this patch is expected to be reverted.

This WA add NO_LPM quirk to the Gen 2 device which directly connected to
the Type-C port of Kinox platform.

BUG=b:245638157
TEST=The Gen 2 device is recognized with either fliped or unflipped cases.
$ cat /sys/bus/usb/devices/2-1/quirks is 0x400 for Gen2 device, otherwise 0x0 for other devices.

Change-Id: I52df568f2f74181561b956523bf1d6bf7244ea3d
Signed-off-by: ronlee <ron.lee@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3958874
Reviewed-by: Sven van Ashbrook <svenva@google.com>
Reviewed-by: Randy Wang <ifanw@google.com>
Reviewed-by: Rajat Jain <rajatja@google.com>
Tested-by: Randy Wang <ifanw@google.com>
Commit-Queue: Randy Wang <ifanw@google.com>