UPSTREAM: wifi: mt76: mt7921s: fix race issue between reset and suspend/resume
It is unexpected that the reset work is running simultaneously with
the suspend or resume context and it is possible that reset work is still
running even after mt7921 is suspended if we don't fix the race issue.

Thus, the suspend procedure should be waiting until the reset is completed
at the beginning and ignore the subsequent the reset requests.

In case there is an error that happens during either suspend or resume
handler, we will schedule a reset task to recover the error before
returning the error code to ensure we can immediately fix the error there.

Fixes: ca74b9b907f9 ("mt76: mt7921s: add reset support")
Co-developed-by: YN Chen <YN.Chen@mediatek.com>
Signed-off-by: YN Chen <YN.Chen@mediatek.com>
Signed-off-by: Sean Wang <sean.wang@mediatek.com>
Signed-off-by: Felix Fietkau <nbd@nbd.name>
(cherry picked from commit e86f10e6809add9132ecc2c6b3184ed59db7ca71)

BUG=b:178754244, b:188614537
TEST=build on cherry

Signed-off-by: Linux Patches Robot <linux-patches-robot@chromeos-missing-patches.google.com.iam.gserviceaccount.com>
Disallow-recycled-builds: test-failures
Change-Id: Id917d3c62fad61c0214e92de10e607b36aff8e55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3958963
Reviewed-by: Jacek Siuda <jsiuda@google.com>
Commit-Queue: Guenter Roeck <groeck@chromium.org>
Reviewed-by: David Ruth <druth@chromium.org>
Tested-by: Guenter Roeck <groeck@chromium.org>