FROMGIT: platform/chrome: cros_ec_debugfs: Set PROBE_PREFER_ASYNCHRONOUS
This driver takes on the order of 40ms to start on some systems. It
shouldn't have many cross-device dependencies to race with, nor racy
access to shared state with other drivers, so this should be a
relatively low risk change.

This driver was pinpointed as part of a survey of top slowest initcalls
(i.e., are built in, and probing synchronously) on a lab of ChromeOS
systems.

Signed-off-by: Brian Norris <briannorris@chromium.org>
Signed-off-by: Tzung-Bi Shih <tzungbi@kernel.org>
Link: https://lore.kernel.org/r/20221101152132.v2.3.Ic9a4f378f73319da323cd55940012fa6b1de24f4@changeid
(cherry picked from commit 692a68ad7f3c568359b9f18d966628856fd34ff3
 git://git.kernel.org/pub/scm/linux/kernel/git/chrome-platform/linux.git for-next)

BUG=b:255668398
TEST=CQ
TEST=platform.BootPerf

Change-Id: Ic9a4f378f73319da323cd55940012fa6b1de24f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3991644
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Benson Leung <bleung@google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Tested-by: Brian Norris <briannorris@chromium.org>