FROMGIT: platform/chrome: cros_ec_lightbar: Set PROBE_PREFER_ASYNCHRONOUS
This driver takes on the order of 15ms to start on some systems. Even on
systems where there is no lightbar support, it can take a few
milliseconds just to probe the EC for support. It shouldn't have many
cross-device dependencies to race with, nor racy access to shared state
with other drivers, so this should be a relatively low risk change.

This driver was pinpointed as part of a survey of top slowest initcalls
(i.e., are built in, and probing synchronously) on a lab of ChromeOS
systems.

Signed-off-by: Brian Norris <briannorris@chromium.org>
Signed-off-by: Tzung-Bi Shih <tzungbi@kernel.org>
Link: https://lore.kernel.org/r/20221101152132.v2.4.I565598102e0bfb03bdf8c090d3bfdf954d026bc5@changeid
(cherry picked from commit 873ab3e886b52ba3d6ade3c5f3e6a0cff0c8cb12
 git://git.kernel.org/pub/scm/linux/kernel/git/chrome-platform/linux.git for-next)

BUG=b:255668398
TEST=CQ
TEST=platform.BootPerf

Change-Id: I565598102e0bfb03bdf8c090d3bfdf954d026bc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3991645
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Benson Leung <bleung@google.com>