FROMGIT: platform/chrome: cros_ec_spi: Set PROBE_PREFER_ASYNCHRONOUS
This driver often takes on the order of 10ms to start, but in some cases
as much as 600ms [1]. It shouldn't have many cross-device dependencies
to race with, nor racy access to shared state with other drivers, so
this should be a relatively low risk change.

This driver was pinpointed as part of a survey of top slowest initcalls
(i.e., are built in, and probing synchronously) on a lab of ChromeOS
systems.

[1] 600ms was especially surprising to me, so I checked a little deeper.
This driver is used to interface with Embedded Controllers besides just
the traditional laptop power-state controller -- it also interfaces with
some fingerprint readers, which may start up in parallel with the
kernel, or which may not even be present on some SKUs, despite having a
node for it. Thus, our time is wasted just timing out talking to it. At
least we can do that without blocking everyone else.

Signed-off-by: Brian Norris <briannorris@chromium.org>
Signed-off-by: Tzung-Bi Shih <tzungbi@kernel.org>
Link: https://lore.kernel.org/r/20221101152132.v2.5.Ia458a69e1d592bfa4f04cde7018bbc7486f91a23@changeid
(cherry picked from commit 015e4b05c377dc5e066e88737bff9990d5ac358d
 git://git.kernel.org/pub/scm/linux/kernel/git/chrome-platform/linux.git for-next)

BUG=b:255668398, b:256936695
TEST=CQ
TEST=platform.BootPerf

Change-Id: Ia458a69e1d592bfa4f04cde7018bbc7486f91a23
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3991647
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>